[Amend] Ignore message, I was looking at the wrong sensor.
Positions kinda work but also very strange. When I'm was right next to the slope it was giving a position of 16.
When using the debugger I can see that block is not being set to the correct values but I don't know why. This commit doesn't change this either. A possibly minor thing I noticed is sometimes the distance is over 16 for some reason