Format and refactor variables to be clearer
Air acceleration is defined as twice the acceleration on ground so I made that relationship explicit. This also has the benefit of not needing to change air acceleration when acceleration changes.

Even though friction is the same as acceleration, I made it into it's own variable. This just makes it clearer when friction is used (self-documenting code.)

As to why I added code braces for a one-liner, I thought having comments around it without the braces was odd.