Refactor finding chunkx/y and tilex/y to their own methods
I was confused on whether to use a Vector2 or not because it would be convenient, but at the same time it uses floats while I'm using ints here, and I wasn't sure whether casting the numerator to float would work properly (it was giving a warning saying there is integer division in a floating point context).

So I just decided to use ints and basically just extract out two methods.

just realised the functions are the exact same though, which is another argument for using a Vector2