Remove most links from SOURCES.md
The Sonic Physics Guide has already been credited as a whole in the README, so these individual links are not necessary.

I also feel that links to official libGDX documentation and other things I used for learning about concepts, while potentially nice reads are not really "sources". I don't think I copy pasted actual code snippets from any of those links. Some The information some of them was referring to didn't even end up getting used at all.

The only real things I found that WERE sources were links to resources and rips of Sonic files. Even then some of them aren't used, but thought I'd put them there anyway for now.

The alsoft fix also wasn't really a source for a change in the actual project, but it is a useful link for fixing an issue that could be present when building and running it, so it could be placed in a more appropriate document later.