Optimise memory allocations by avoiding unnecessary instantiation
I profiled the game, and it turned out that A LOT of memory was being allocated in this specific part of the code - evidently due to the creation of a new Color object.

One instantiation isn't a problem in and of itself, but this method is called many, many, many times per second (for almost every column of pixels in each tile, this line will be run, and that's per frame).

Thankfully the batch has a setColor method that just takes the RGBA values directly instead of a Color - so I don't need to create a new Color.

The allocation size dropped from around 600MB to 20MB after this change in a very similar run-through of the game.

Of course, the Colors will be garbage collected so memory *usage* might not be much different, but if we can avoid this allocation there is no reason not to.

I looked through the setColor code and it calls color.set(). The methods are very similar with and without creating a new Color so there shouldn't be much of a performance difference.