Make Tiles Immutable
I was considering keeping attributes non-final (so that there could be "unsafe" methods which modify them) but I decided not to do so because right now I don't really see how such methods would be beneficial. Hypothetical "unsafe" methods which modify a tile would modify the same tile in every chunk (because each reference points to the same tile object) which would change collision of a level in potentially unexpected ways.

I still added getters and setters, however, since I don't think speed is much of an issue and this makes the class more flexible to be potentially changed in the future (e.g. if I actually decide that I DON'T want the attributes to be final).

All attributes have been kept final, but it is most important that the arrays are, since they are not actually immutable. The getter only returns an element of the array (which is "defensively" copied so changing it outside won't change the inside array) which prevents the underlying array being exposed.

To make the attributes final, I had to instantiate the attributes for the empty tile. This is good practice anyway though, because it prevents potential errors caused if empty tiles are treated like non-empty ones unintentionally (which I guess you could consider a bad thing too).