Override Tile equals() and hashcode()
Technically the tiles are statically initialized and placed, so == should work, but I thought overriding equals() would be more robust and is likely to be needed anyway in the future.

Got IntelliJ to do it for me, the only thing to think about was whether to use instanceof or getClass as part of the equals method. I decided on getClass, because I'm only really worried about the Tile class and am not planning for any subclasses, at least right now. So really, it didn't matter which approach I chose here.

See https://stackoverflow.com/questions/596462/any-reason-to-prefer-getclass-over-instanceof-when-generating-equals