Use Optional instead of returning null
Using Optional makes it clearer that the method can return "null" and forces the user to deal with it. I don't particularly care about the possible minor performance loss.

For the floorSensors use, I hoisted the null check out because it seemed every check required winningSensor to be null.
There is one semantic difference after the optional change inside the first if statement, because previously it would enter that statement and execute the else part, setting isGrounded to be false. But I think it's actually BETTER that it ISN'T set, since if it's null the player is likely to still be on the ground.

While rewriting JavaDoc and I retested the floorSensors conundrum which originally made me return null. When I returned sensorA instead of null, the player would jitter when moving between an upwards facing slope and ground. When returning sensorB, they would jitter when moving between the ground and downwards facing slope. Concluded that it seems that it IS necessary to do some sort of null thing in this method.

(I edited the level to add that extra piece of ground to figure this out).

As for the getChunk optional change, it was more self-explanatory (apart from me inverting the check in checkEmpty because I didn't like having two different isEmpty methods). Optional can be empty and the Tile can be empty.