Rename TILE_MAP to INSTANCE
I found this naming convention for enum singletons from stack overflow (https://stackoverflow.com/a/9850559) and find it pertinent for TileMap.

TileMap.TILE_MAP was a very confusing name, and INSTANCE makes the code using it make a lot more sense and the purpose of using an enum clearer (easy implementation of singleton). It still is far from ideal though, and I suspect the class is going to change a lot more from here.

Also reordered the if statement comparison checks to be consistent with how it's done elsewhere.

Note I haven't been consistently updating the commented out obsolete code. I may just end up removing it entirely, that's bound to happen sooner or later.