Use getChunk method in drawChunkTexture and make getTexture return Optional
I saw that this was one of the few areas where map was being accessed directly in a different class. The only other two instances were when checking the length of the map, which I will deal with later.

The trouble with the getChunk method is I had to deal with the optionals, but that of course is the point. Since I realised Texture could also be null if the Chunk is empty, instead of checking if empty I decided to make the chunk method return an Optional to bulletproof the code. This meant I had to unwrap two optionals to get the texture though.

Before I said I wasn't concerned about performance, but I sort of retract that statement here. Similar to the previous situation with tons of memory being allocated in another draw method, this drawChunkTexture method is being called many times per frame and hundreds of times per second. It is not a great idea to instantiate an object in this loop. Declaring references to existing objects is fine, though.