Use Optional as Chunk field and return empty chunk instead of Optional in getChunk
Returning an empty chunk was simple, and the only downside is that the return type doesn't force the user of the method to check that the Chunk is empty. However, the Optional<Texture> mandates the Texture being checked anyway, so null issues are unlikely. Not to mention that the array could contain empty chunks anyway, so users would have to check anyway. But I added a javadoc to state that the Chunk could be empty.

As for the Optional field, I was unsure on whether to do this, because the Java language designers only intended Optional to be used as a return value (and it isn't serializable which may or may not be an issue later). I was thinking of just using null, but I decided the benefit of optional outweighed the downsides. I still didn't want to use it as a return value though because it would result in lots of duplicate calls to of() every frame, therefore creating lots of optional objects unnecessarily when it can just be done once per Chunk.

I skipped the isPresent check in GameScreen because I'm already checking that the chunk is empty.

As an extra change, I inverted the if statement checks in the getChunk method so that the edge case is checked first. It's subjective but I think the layout makes it clearer that that is an edge case. I also added logging because ideally the user shouldn't look up chunks at invalid ranges anyway.

and I renamed eChunk to emptyChunk because before it was a bit ambiguous - e could stand for something entirely different.