Remove debug message from getChunk
Ironically it was causing an ArrayIndexOutOfBoundsException as the if statement condition was triggered, so it would go into that code but the debug message required getting the length of map[chunkX]. So if chunkX was out of range, it would throw an error. There are obviously ways this could be fixed (e.g. by using multiple if statements and not always logging map[chunkX].length) but I don't really like having this message repeat all the time anyway. I know about this issue.