TEST: Commented out setting the batch back to white to keep the tint.
Because I used the sprite's draw method for Sonic, the color wasn't respected there. But it was for the chunks, and it can have a nice effect sometimes. But I realised that other times it can make the tiles harder to see with some tints (mainly green ones) because it blends into the background.