Remove set and get position vector methods from Sensor
Problem with getPositionVector is that libGDX Vector2 objects are mutable. Therefore the position vector could easily be "corrupted" or mutated outside of this class just by using the getter and then calling methods on the returned vector. This isn't great practice, and the getXPosition and yPosition methods are preferable because they return floats directly, which are of course passed by value since they are primitives so you could think of them as being defensively copied.

Similarly for the setter, the caller to that method could retain a reference to the vector which was passed in and then mutate it afterwards.

Obviously I wouldn't be intentionally doing these things anyway, but that doesn't mean I couldn't do them unintentionally.